% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneNetwork.R
\docType{data}
\name{geneNetwork}
\alias{geneNetwork}
\title{The gene network used in simulation studies}
\format{
A binary adjacency matrix 0/1
}
\usage{
data(geneNetwork)
}
\description{
The gene network used in simulation studies which is extracted from real data network, represented by a binary adjacency matrix 0/1.
}
\details{
In the simulation studies, the gene network is a subnetwork extracted based on real gene network. It contains 776 gene network nodes in total
}
\keyword{datasets}
