% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingLocations.R
\docType{data}
\name{missingLocations}
\alias{missingLocations}
\title{The missing locations of gene vector in the simulated dataset}
\format{
A vector of length 157, representing the missing location in the gene vector. This is generated with a 20\% missingness mimicking the real data missing rate
}
\usage{
data(missingLocations)
}
\description{
The simulated missing locations
}
\details{
The missing locations are randomly generated among 'non-hub' genes for the simulated gene network (776 gene nodes in total)
}
\keyword{datasets}
