% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryClassLabels.R
\name{summaryClassLabels}
\alias{summaryClassLabels}
\title{{Local FDR method for summarizing class labels}}
\usage{
summaryClassLabels(net,test.stat,zTrack,method=c("MajorVote","LocfdrControl"),
plot=FALSE,nburn=50,nskip=2,locfdr.alpha=0.2)
}
\arguments{
\item{net}{The binary adjacent matrix with 0/1 indicating "connected" or "not directly connected}

\item{test.stat}{A vector represents the observed test statistics. Missing values are represented as NAs}

\item{zTrack}{The trace matrix for class labels. It is a num_of_genes by num_of_iterations matrix}

\item{method}{A stirng. The name of summarizing method, can be "MajorVote" or "LocfdrControl"}

\item{plot}{Logical. The default is FALSE, otherwise the plot will be drawn in the end}

\item{nburn}{The number of burn-in iterations. The default is 50}

\item{nskip}{The thinning interval. The default is 2}

\item{locfdr.alpha}{The maximum tolerated rate of falsely discovered genes. The default is 0.2}
}
\value{
A vector where each element is the summarized class label, the values can be 1,2 or 3 representing "down-regulated gene", "null gene /not differentially expressed gene" or "up-regulated gene"
}
\description{
The final class labels can be decided either by method "Major vote", or "under local fdr control".
}
\examples{
\dontrun{
data(geneNetwork)
data(testStatistics)
res=BNC(net=geneNetwork,test.stat=testStatistics,niter=100,na.action="NN")
classLabelEst=summaryClassLabels(net=geneNetwork,test.stat=testStatistics,
zTrack=res$zTrack,method="MajorVote",plot=FALSE)
print(table(classLabelEst))
}
}
