% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\docType{data}
\name{testStatistics}
\alias{testStatistics}
\title{Test statistics of genes}
\format{
A vector of the test statistics for each gene node. The length is the total number of gene nodes in the network
}
\usage{
data(testStatistics)
}
\description{
The simulated test statistics of gene nodes. With 10\% missing rate that is randomly generated
}
\details{
Firstly, we simulate the underlying true gene class labels by 25\% down-regulate, 50\% null (not differentially expressed) and 25\% up-regulate. The pre-defined distribution for each regulation group is the normal distributions. The down-regulated genes are distributed of a normal distribution with mean=-0.5 and sd=0.2, the null genes are distributed of a normal distribution with mean=0 and sd=0.2, the up-regulated genes are distributed of a normal distribution with mean=0.5 and sd=0.2. Secondly, we randomly knock out 10\% genes as missing nodes
}
\keyword{datasets}
