% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNCmainfun.R
\name{TransZ}
\alias{TransZ}
\title{{TransZ}}
\usage{
TransZ(p)
}
\arguments{
\item{p}{The probability of a down-regulated gene jumping to null gene}
}
\value{
A 3*3 transition matrix for Z
}
\description{
Compute a special transition matrix of Z: symmetric, down-regulated genes are not able to jump directly to up-regulated genes; a probability of a down-regulated gene jumping to null gene equals the probability of an up-regulated gene jumping to a null gene.
}
\keyword{internal}
